requirements.txt
numpy~=1.22.4
scikit-image~=0.18.3
opencv-python~=4.6.0.66
pillow~=9.1.1
tensorflow~=2.9.0
ipython~=8.4.0
scikit-learn~=1.0.2